Refactor: structure of static/assets
