+++ src/components/ModalPopup/Body/BodyCallTaxi.tsx
@@ -5,9 +5,9 @@ import LinkCallTaxi from "components/Link/LinkCallTaxi";
 import theme from "tools/theme";
 
 import LocationOnRoundedIcon from "@mui/icons-material/LocationOnRounded";
-import { ReactComponent as KakaoTaxiLogo } from "static/assets/KakaoTaxiLogo.svg";
-import TmoneyOndaLogo from "static/assets/TmoneyOndaLogo.png";
-import { ReactComponent as UTLogo } from "static/assets/UTLogo.svg";
+import { ReactComponent as KakaoTaxiLogo } from "static/assets/serviceLogos/KakaoTaxiLogo.svg";
+import TmoneyOndaLogo from "static/assets/serviceLogos/TmoneyOndaLogo.png";
+import { ReactComponent as UTLogo } from "static/assets/serviceLogos/UTLogo.svg";
 
 export type BodyCallTaxiProps = {
   roomInfo: Room;
