+++ src/components/ModalPopup/ModalChatPayment.tsx
@@ -19,8 +19,8 @@ import theme from "tools/theme";
 import CheckRoundedIcon from "@mui/icons-material/CheckRounded";
 import ContentCopyIcon from "@mui/icons-material/ContentCopy";
 import LocalAtmRoundedIcon from "@mui/icons-material/LocalAtmRounded";
-import { ReactComponent as KakaoPayLogo } from "static/assets/KakaoPayLogo.svg";
-import { ReactComponent as TossLogo } from "static/assets/TossLogo.svg";
+import { ReactComponent as KakaoPayLogo } from "static/assets/serviceLogos/KakaoPayLogo.svg";
+import { ReactComponent as TossLogo } from "static/assets/serviceLogos/TossLogo.svg";
 
 type ModalChatSettlementProps = Omit<
   Parameters<typeof Modal>[0],
