+++ src/components/ModalPopup/ModalPrivacyPolicy.tsx
@@ -6,7 +6,7 @@ import BodyPrivacyPolicy from "./Body/BodyPrivacyPolicy";
 
 import theme from "tools/theme";
 
-import { ReactComponent as TaxiLogo } from "static/assets/TaxiLogo.svg";
+import { ReactComponent as TaxiLogo } from "static/assets/sparcsLogos/TaxiLogo.svg";
 
 type PopupPrivacyPolicyProps = {
   isOpen: boolean;
