+++ src/components/ModalPopup/ModalTerms.tsx
@@ -19,7 +19,7 @@ import { useSetRecoilState } from "recoil";
 
 import theme from "tools/theme";
 
-import { ReactComponent as TaxiLogo } from "static/assets/TaxiLogo.svg";
+import { ReactComponent as TaxiLogo } from "static/assets/sparcsLogos/TaxiLogo.svg";
 
 type ModalTermsProps = {
   isOpen: boolean;
